+++ Carbon/Collections/Dictionary.lua
@@ -156,9 +156,6 @@ Dictionary.Copy = ShallowCopy
 
 	Performs a self-reference fixing deep copy from one table into another.
 	Handles self-references properly.
-
-	FIXME: datawise = 1 triggers a shallow datawise copy (only first layer is datawise)
-	This could be more elegant.
 }]]
 function Dictionary.DeepCopy(self, to, datawise, map)
 	to = to or Dictionary:New()
@@ -173,10 +170,13 @@ function Dictionary.DeepCopy(self, to, datawise, map)
 				local copy = (not datawise) and (value.Copy or value.DeepCopy or value.ShallowCopy)
 
 				if (copy) then
-					map[value] = copy(value, nil, datawise ~= 1 and datawise, map)
+					map[value] = {}
+					copy(value, map[value], datawise, map)
 				elseif (t == "table") then
 					map[value] = {}
-					Dictionary.DeepCopy(value, map[value], datawise ~= 1 and datawise, map)
+					Dictionary.DeepCopy(value, map[value], datawise, map)
+				else
+					map[value] = value
 				end
 			end