+++ changes.md
@@ -3,6 +3,7 @@
 ![1.1.0](https://img.shields.io/badge/1.1.0-in_development-orange.svg?style=flat-square)
 - Improved Carbide parsing rigidity
 - Improved overall math performance
+- Added fat-arrow lambda syntax (`=>`)
 - Added `Bench` class for benchmarking
 - Added custom allocators in the form of the `Allocator` attribute
 - Added new Quaternion methods: